refactor api for working with Resources
WebFinger resources can simply be treated as URLs with some extra logic
surrounding extracting the host for acct and mailto URLs.  This now
supports all types of URLs, not just email addresses, however if an
email-like identifier is provided, it is treated as an acct URI.  Full
tests for the Resource type have been added as well.

!! This is a breaking change.