remove legacy fallback
Since that was the only thing using XRD, this removes the XRD package as
well.  Technically, it's still valid in the latest WebFinger spec to
request a non-standard content-type, but in practice no one does, so
there's little point in keeping support around.

This also makes the following changes to bring things in line with the
latest draft spec:
 - add support for application/jrd+json in addition to application/json
 - no longer require that the JRD Subject match the requested resource